Add Fairring to PARAM-comms
Summary: In order to use PARAM-comms to benchmark Fairring without needing local patches or ephemeral fbpkgs it would be ideal to add support for it upstream. Here I do so, by making it an "optional" dependency, just like UCC: we only attempt to import and use it if the user explicitly asks for it. In all other cases there's no difference in behavior.

Reviewed By: kingchc

Differential Revision: D33659403

fbshipit-source-id: 538a6e54886ea63a2cfca36247b3ce954a865f89