Invoke all-gather with tensors from same storage
Summary: This change is a bit silly and it's only needed to work around an idiosyncrasy of Fairring: to prevent the user from accidentally triggering a memory copy, Fairring will refuse to operate on reduce-scatter or all-gather tensors that do not come from a single contiguous storage (as that's the format that NCCL needs at a low level). The only place in PARAM where this isn't already happening is when aggregating stats. The change is very simple and should not have any other visible side-effect.

Reviewed By: kingchc

Differential Revision: D33764338

fbshipit-source-id: 8528b53e97db1e936482f8a3750529afd2b9faa0