avoid double shrinking when replaying a2av with smaller scales
Summary:
**WHY**
In auto-shrink, we reduce message size to fit current process count. However, we shrink it twice mistakenly (one during warm-up and one during real run), then the message sizes and splits do match up for Allttoall that leads to runtime errors.

The root cause is that we share the same object to store comms info. This patch makes a copy of comms info during warm-up to avoid it gets shrink twice.

Reviewed By: bryanmr

Differential Revision: D34771731

fbshipit-source-id: 7201990500b383f332bef884a7cbb3521d19de47