Updating the arguments to accept DLRM like MLP config.
Summary:
Updating the linear benchmark to accept the layer configs similar to DLRM, as a dash separated int string. This simplifies benchmarking components and specifying such benchmarks with vendors.

To maintain the current arguments, the `layers_size` argument is added which takes precedence. Either way of specifying the config will first put the config in `laters_size` array, on which the rest of the functions operate.

Reviewed By: nrsatish

Differential Revision: D36029864

fbshipit-source-id: 1dde685237d980d9d00e67db49cb572647a7a669