Improve wait operation to better support asynchronous testing.
Summary:
Each collective is associated with a request_id. When there is a wait operation in the trace, we need to wait for the collective operation associated with that request_id, and the collectives prior to that one to be done before moving on to issuing the next collective operation.

Wait is a noop for now because CUDA will go OOM for nonblocking runs if we don't post collectives immediately.

Reviewed By: Pavani-Panakanti

Differential Revision: D37028726

fbshipit-source-id: 19d1b6abe98ee98a43bac0c1542b9c9f2fd3614f