Add reduce_scatter_v to collective benchmarking
Summary: This change adds benchmarking for `dist.reduce_scatter_v`. Input split sizes are provided using the --i or --in_split flag as a list of input split sizes per rank delimited by commas. If the --i flag is provided, the --b and --e flags are ignored, and the buffersize is determined according to the sum of input splits. Also adds an out_split flag to allow uneven splits for all_to_allv.

Reviewed By: kingchc

Differential Revision: D37640725

fbshipit-source-id: 5d367f764821d578a3112edc6a3ebdcfeea87de1