Add all_gather_v to collective benchmarking
Summary: This change adds benchmarking for `dist._all_gather_v`. Output split sizes are provided using the --o or --out_split flag as a list of output split sizes per rank delimited by commas. If the --o flag is provided, the --b and --e flags are ignored, and the buffersize is determined according to the sum of output splits.

Reviewed By: kingchc

Differential Revision: D38341688

fbshipit-source-id: c5bea78623d33fc3f6fdf71c5ab50574c91e972f