Rename num-compute to num-compute-per-iteration
Summary:
Clears the way for a more general version of the benchmark which could have more than one compute in an iteration independent of the number of collectives.

This is backward compatible with the existing --num-compute argument as it only adds an alias and renames the internal variables to be consistent.

Reviewed By: kingchc

Differential Revision: D40229127

fbshipit-source-id: 49b36da77541db5f57ac8b748d85303a4729af51