Add batched embedding and pooled alltoall
Summary:
The current param benchmarks include basic collective and compute overlaps, but there are some specific types of operations that are more common in production workloads. Alltoall + embedding is a common workload, but we don't currently have a good way to measure the performance of batching some of the embedding tables together for the embedding lookup and then communicating those tables through alltoall (as opposed to completing the entire embedding lookup and the entire alltoall at once).

This adds the ability to batch together embedding tables with `--num-emb-tables-batched`. If this flag is used with an alltoall collective, the alltoall will take that batching into account and switch to using `alltoall_pooled` from the `extend_distributed` package.

Reviewed By: kingchc

Differential Revision: D40352875

fbshipit-source-id: 1afed5399b55ecce616d03d0d2af450e4c1c514c