Update param_comm benchmarks to accept a comma-separated list of data-types as argument. Also append the timestamp to the log file name
Summary:
Add a new argument --data-types for ParamCommBench which accepts a list of comma-separated ddata-types.
Update the ParamCommBench to run iteratively once for each data-type received in args.data_types. Append the param-type to report containing benchmark-Times for each collective to accurately depict which datatype was the collective run for.

To maintain backward compatibility, both --data-type and --data-types args can be used interchangeably. If not provided, ParamCommBench will use default data-type float32 as data-type for benchmarking purposes

Reviewed By: kingchc

Differential Revision: D42176780

fbshipit-source-id: 09c768c2b3caff6dd81c85b5571e14225baafa05