optionally use TCP store to bcast trace file
Summary: This patch adds --disable-parallel-read option. It is false by default meaning every rank will read the trace file (same as original behavior). When it is set to true, only rank 0 will load the trace and it will broadcast to others. This must be used only when --use-one-trace is True.

Reviewed By: kingchc

Differential Revision: D45103723

fbshipit-source-id: 91b6ef7bc003f374c52e1a27294d7b70da2464b0