always specify device_ids in barrier
Summary:
NCCL PG internally checks if a device has been selected in barrier, otherwise it reports warning below.
```
Rank 0 using GPU 0 to perform barrier as devices used by this process are currently unknown. This can potentially cause a hang if this rank to GPU mapping is incorrect.Specify device_ids in barrier() to force use of a particular device.
```

To avoid the warning, always pass device_ids to barrier in param.

Reviewed By: kingchc

Differential Revision: D45213972

fbshipit-source-id: 8314f19ad475917d87f775386fcf5e268fc8c958