use a smaller scaleFactor to avoid overflow
Summary:
When used previous large scaleFactor (numElementsOut*numElementsOut) which may exceeds the range of long, using it to divide random tensor value can cause "overflow when unpacking long" error.

This patch uses world_size as the scaleFactor which is always in integer range. Note that, scaleFactor is used only for random tensor value when dcheck is disabled, thus no impact to correctness check.

Reviewed By: kingchc

Differential Revision: D47001230

fbshipit-source-id: 0f8cc0b8c348f858641ee8234f472d043ad77589