report msg size in bytes in replay mode
Summary:
In reply mode, we print the sizes of input tensor and output tensor in bytes at end of run. However, the recorded values are num of elements not bytes. This patch fixes it.

Since `size` is unclear whether in bytes or in num of elements, this patch also renamed the recording attributes sizes->bytes. E.g., collInMsgSizes->collInMsgBytes

This patch also adds NumElemsIn, NumElemsOut, Dtype in per-replay log.

Reviewed By: kingchc

Differential Revision: D47316759

fbshipit-source-id: 00a6a835e9aaaaef7dab19bbe417f7dec4b1615d