Add validation check for input and trace_path arguments
Summary:
Previously, 'eg_replay.py' did not handle a scenario where the 'input' and 'trace_path' arguments were both not provided. This led to a TypeError when the program attempted to iterate over 'NoneType'.

The error was as follows:
```
Traceback (most recent call last):
...
  File ".../eg_replay.py", line 208, in initBench
    if "://" in self.args.input:
TypeError: argument of type 'NoneType' is not iterable
```
In this commit, I have addressed this scenario. Now, if both 'input' and 'trace_path' arguments are not provided, the program will print a help message and exit.

Reviewed By: briancoutinho

Differential Revision: D47670146

fbshipit-source-id: d1f1435c1bd294821e3bdf8bf28628da6169c3a6