add --multi-comms option
Summary:
It allows user to run multiple concurrent collectives each with different communicator. E.g., set --multi-comms 8 with ppn 8 would create 8 communicators, each with 1 GPU per node.

It supports communicator division on the ppn dimention. We may extend it in future to divide on the node dimention.

Limitation: it cannot be used together with --overlap-pair-pgs, because both use the same code to generate groups. It will report error if user enables them together.

Reviewed By: steinbrecher

Differential Revision: D52855337

fbshipit-source-id: 27ad0d4007c48eb616195d07491b59cf58962dbb