Sync GPU device for discrete benchmark (#97)
Summary:
Pull Request resolved: https://github.com/facebookresearch/param/pull/97

This diff aims to synchronize the GPU device for discrete benchmark. In discrete benchmark, the op run time is measure after the op runs. However, on device except cuda:0, the run is a-synchronized, the measured time is only the host runtime.

Added torch.cuda.synchronize call to wait until the op run finishes on GPU.

Differential Revision: D54907227