Fixed the bug to access tensor stride (#150)
Summary:
Pull Request resolved: https://github.com/facebookresearch/param/pull/150

There is a bug to access tensor stride. The current implementation uses the index of the input tensor to access node.tensor_strides. However  node.tensor_strides is for all input variables including non tensor variables. So the fix is to get the list of tensor stride, then uses the index of the input tensor to access the list of tensor stride.

Reviewed By: briancoutinho

Differential Revision: D60642585