Initial param integration with Geordi
Summary:
Example of the integration.

Adds a new envrionment (for now) that uses the param benchmarker instead of the builtin benchmarker. The current output from param is not sufficent to return benchmarking results so we don't output a valid benchmark result or upload the result to Warehouse.

We'll land this first to test out param and then start work on the dataswarm pipeline

Reviewed By: shengfukevin

Differential Revision: D63269468

fbshipit-source-id: fa7d2152a2108c6959dcc8b43bafbfb2c7e70c2b