Share tensors between comp replay and comms replay
Summary:
Share tensors between comp replay and comms replay

When run full replay in et_replay, compute replay and comms replay manage the tensor allocation separately. So some tensors are double allocated, this leads to the full replay of Llama4 70B out of memory.

This DIFF is to fix it by allocating tensors in comp replay and passes them to comms replay.

Reviewed By: sanrise

Differential Revision: D67353163