Support replay config file
Summary:
X-link: https://github.com/pytorch-labs/chakra_replay/pull/8

Support replay config file

Changed command line option "skip-node-file" to "replay-config" and "update-skip-node-file" to "update-replay-config".

The skip nodes are not included in replay config file.

The following is one example of a config file:

{
    "import modules": [
        "fbgemm_gpu.split_table_batched_embeddings_ops_training"
    ],
    "skip nodes": {
        "aten::_foreach_copy_": "Failed to build function",
        "aten::_foreach_norm": "Failed to build function",
        "aten::_foreach_div_": "Failed to build function",
        "aten::_foreach_mul_": "Failed to build function",
        "aten::_foreach_add_": "Failed to build function"
    }
}

"import modules" include the list of third party modules that need be imported to et-replay to make custom ops available in replay.

"skip nodes" defines the list of ops that need be skipped in replay. The first entry is the node name and the second entry is the reason why it is skipped.

Reviewed By: c-odrin

Differential Revision: D78688077

fbshipit-source-id: 90d9a2da7fca8352be6b6759aacb0dc97c48a805