Removing Unused import(s) in fbsource/fbcode/param_bench/train/compute/python/workloads/pytorch/native_basic_ops.py
Summary:
Removed the unused import `from typing import Dict` from the file. The file uses the built-in `dict` type annotation (Python 3.9+) instead of the deprecated `typing.Dict`, so this import is not needed.

This change improves code quality by removing an unused import, which was flagged by flake8 rule F401.
 ---
> Generated by [RACER](https://www.internalfb.com/wiki/RACER_(Risk-Aware_Code_Editing_and_Refactoring)/), powered by [Confucius](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=0dd7bbf6-b059-11f0-a909-0a44f5fef2c5&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=0dd7bbf6-b059-11f0-a909-0a44f5fef2c5&tab=Trace)

Differential Revision: D85381933

fbshipit-source-id: edaac99da6d07aa70427e3167e37d060e88705b1